Push Go source instead of a binary
This makes the uptimer binary all you need to run uptimer.
There is no longer a requirement to have a valid Go toolchain
insalled on the machine running uptimer.

To do this, we write a CF manifest and a main.go source file
to a temp directory. As a result, the compilation of the apps
for pushability and syslog sink tests is offloaded to the
platform via the Go buildpack.

For the syslog sink test app, which has dependencies on the
rfc5424 package, we vendor all of the dependencies into a single
file.  This was performed by concatenating all of the source,
removing spurious package and import clauses, and embedding
the result into a string constant.

    cd syslogSink
    cat vendor/code.cloudfoundry.org/rfc5424/marshal.go \
        vendor/code.cloudfoundry.org/rfc5424/message.go \
        vendor/code.cloudfoundry.org/rfc5424/stream.go \
        vendor/code.cloudfoundry.org/rfc5424/unmarshal.go \
        vendor/code.cloudfoundry.org/rfc5424/rfc5424.go \
        >> app.go

Updates #22