CI: stop following develop->main release flow
We've only been merging to main for 3 years now. CI should reflect that
reality by ignoring the develop branch. Additionally, since we don't cut
releases for uptimer and instead just use the latest, we don't need a
ship-it job anymore as there's no tagging or release creation to be
done.