Fix cfWorkflow tests
Fixes the cfWorkflow tests, which have occasionally failed consistently,
and occasionally worked consistently.

Whenever the tests were run with go >= 1.20.0 the log streaming test
would fail, and it was unclear why. The actual `exec.Cmd` struct and the
expected `exec.Cmd` struct output by `cfCmdGenerator.StreamLogs`
appeared to be exactly the same when printed out.

Upon further inspection, it seems that in go 1.20, a new `Cancel` field
was added to the `exec.Cmd` struct. For some reason, `gomega.Equal`,
which uses `reflect.DeepEqual` under the hood, does not judge the
`Cancel` field on two seemingly identical `exec.Cmd` structs to be
equal. Also and `cfCmdGenerator.StreamLogs` is the only method that
generates an `exec.Cmd` struct with a context, which is how the `Cancel`
field becomes non-nil.

This change switches the comparison to use BeComparableTo, which uses
`go-cmp` under the hood, and allows for extra options to be passed in.
Using these options I turned off comparing the `Cancel` field. However,
in order to make it work I also apparently needed to ignore unexported
fields (just the `ctx` field, for now), and the `Err` field. Not great,
but should be fine because we really only care about the `Args`, `Path`,
and `Env` fields.